feat: separate type defs from functionality
