+++ src/account/balance/index.ts
@@ -1,83 +1,2 @@
-import { EtherscanBaseUrl } from "etherscan/constants";
-
-import type {
-  EtherscanAction,
-  EtherscanActionCall,
-  EtherscanParams,
-  EtherscanResponse,
-  EtherscanResult,
-} from "etherscan/types";
-
-import type { AddressParam, TagParam } from "etherscan/types/param";
-
-export type BalanceParam = AddressParam | TagParam;
-
-export type BalanceParams = EtherscanParams<BalanceParam>;
-
-export type BalanceResult = EtherscanResult<bigint>;
-
-export type BalanceResponse = EtherscanResponse<BalanceResult>;
-
-export type BalanceActionCall = EtherscanActionCall<
-  BalanceParams,
-  BalanceResponse
->;
-
-export type BalanceActionName = "balance";
-
-export type BalanceAction = EtherscanAction<
-  BalanceActionName,
-  BalanceActionCall
->;
-
-export const balance: BalanceActionCall = async (
-  baseUrl: EtherscanBaseUrl,
-  apiKey: string,
-  params: BalanceParams,
-): Promise<BalanceResponse> => {
-  const queryParamsString = `module=account&action=balance&address=${params.address}&tag=${params.tag}`;
-  const url = `${baseUrl}/?${queryParamsString}&apikey=${apiKey}`;
-
-  const response = await fetch(url);
-  const balanceResponse = await response.json();
-
-  return {
-    ...balanceResponse,
-    result: BigInt(balanceResponse.result as string),
-  };
-};
-
-if (import.meta.vitest !== undefined) {
-  const { it, expect, describe } = import.meta.vitest;
-
-  describe("balance", () => {
-    const apiKey = process.env.ETHERSCAN_API_KEY ?? "";
-    const balanceParams = {
-      address: process.env.SEPOLIA_ADDRESS ?? "",
-      tag: "latest",
-    };
-
-    console.log(apiKey);
-    console.log(balanceParams.address);
-
-    it("should return a bigint result", async () => {
-      const { result } = await balance(
-        EtherscanBaseUrl.Sepolia,
-        apiKey,
-        balanceParams,
-      );
-
-      expect(result).toBeTypeOf("bigint");
-    });
-
-    it("should return a balance greater or equal than zero", async () => {
-      const { result } = await balance(
-        EtherscanBaseUrl.Sepolia,
-        apiKey,
-        balanceParams,
-      );
-
-      expect(result).toBeGreaterThanOrEqual(0);
-    });
-  });
-}
+export * from "./call";
+export * from "./types";
